+++ web/next.config.js
@@ -10,7 +10,7 @@ const withPWA = require('next-pwa')({
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  output: 'export',
+  // output: 'export',
   distDir: process.env.NODE_ENV === 'development' ? 'out' : 'dist',
   images: { unoptimized: true },
   webpack: (config, { defaultLoaders }) => {
@@ -21,11 +21,11 @@ const nextConfig = {
     config.resolve.modules.push(path.resolve(`./`))
 
     return config
-  }
+  },
   // for local development - (\\d{1,}) is for port number
-  // async rewrites() {
-  //   return [{ source: '/api/parse', destination: 'http://localhost:8080/api/parse' }]
-  // }
+  async rewrites() {
+    return [{ source: '/api/parse', destination: 'http://localhost:8080/api/parse' }]
+  }
 }
 
 const withALL = (nextConfig = {}) => withContentlayer(withPWA(nextConfig))
