+++ web/package.json
@@ -17,22 +17,22 @@
     "@radix-ui/react-dropdown-menu": "^2.0.5",
     "@radix-ui/react-scroll-area": "^1.0.4",
     "@radix-ui/react-separator": "^1.0.3",
-    "@types/node": "20.4.8",
-    "@types/react": "18.2.18",
+    "@types/node": "20.5.1",
+    "@types/react": "18.2.20",
     "@types/react-dom": "18.2.7",
-    "autoprefixer": "10.4.14",
+    "autoprefixer": "10.4.15",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
     "concurrently": "^8.2.0",
     "contentlayer": "^0.3.4",
-    "framer-motion": "^10.15.1",
+    "framer-motion": "^10.16.0",
     "highlight.js": "^11.8.0",
     "mdast-util-toc": "^7.0.0",
-    "next": "13.4.13",
+    "next": "13.4.19",
     "next-contentlayer": "^0.3.4",
     "next-pwa": "^5.6.0",
     "next-themes": "^0.2.1",
-    "postcss": "8.4.27",
+    "postcss": "8.4.28",
     "postcss-focus-visible": "^9.0.0",
     "postcss-import": "^15.1.0",
     "react": "18.2.0",
@@ -49,6 +49,7 @@
     "shiki": "^0.14.3",
     "tailwind-merge": "^1.14.0",
     "tailwindcss": "3.3.3",
+    "typeit-react": "^2.6.4",
     "typewriter-react": "^1.0.1",
     "unist-builder": "^4.0.0",
     "unist-util-visit": "^5.0.0",
@@ -57,17 +58,17 @@
   },
   "devDependencies": {
     "@types/unist": "^3.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.3.0",
-    "@typescript-eslint/parser": "^6.3.0",
-    "eslint": "8.46.0",
-    "eslint-config-next": "13.4.13",
+    "@typescript-eslint/eslint-plugin": "^6.4.0",
+    "@typescript-eslint/parser": "^6.4.0",
+    "eslint": "8.47.0",
+    "eslint-config-next": "13.4.19",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-react": "^7.33.1",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^3.0.1",
-    "prettier-plugin-tailwindcss": "^0.4.1",
+    "prettier": "^3.0.2",
+    "prettier-plugin-tailwindcss": "^0.5.3",
     "typescript": "5.1.6"
   }
 }
